+++ src/main/java/com/tracebucket/moo/service/IContactService.java
@@ -2,6 +2,7 @@ package com.tracebucket.moo.service;
 
 import com.tracebucket.moo.domain.Contact;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -9,10 +10,10 @@ import java.util.List;
  */
 public interface IContactService {
     public List<Contact> findAll();
-    public Contact findOne(Long id);
+    public Contact findOne(BigInteger id);
     public Contact create(Contact contact);
     public Contact update(Contact contact);
-    public Boolean delete(Long id);
+    public Boolean delete(BigInteger id);
     public Contact addContact(Contact contact);
     public List<Contact> findByLastName(String lastName);
